% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{api_connect}
\alias{api_connect}
\title{Fonction pour établir la connexion à l'API}
\usage{
api_connect(api_key, api_secret)
}
\arguments{
\item{api_key}{La clé d'API utilisée pour l'authentification.}

\item{api_secret}{Le secret d'API utilisé pour l'authentification.}
}
\value{
Une liste contenant l'URL de base de l'API et les en-têtes d'authentification.
}
\description{
Cette fonction établit une connexion à l'API en utilisant les clés d'API fournies.
}
\examples{
# Utilisation de la fonction pour établir une connexion à l'API
api_key <- "8c2cf1af-c970-47dd-a3fb-9841b2e9e8ed"
api_secret <- "pot_SvtyFegiqK2YckFNfAXPgjdNtVrM8BdW_jKd-zo"
connection_info <- api_connect(api_key = api_key, api_secret = api_secret)

}
