% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_news}
\alias{get_news}
\title{Fonction pour récupérer les news par les endpoints}
\usage{
get_news(
  connect_info,
  start_date,
  end_date,
  query = NULL,
  source = NULL,
  category = NULL,
  country = NULL
)
}
\arguments{
\item{connect_info}{Une liste contenant les informations d'authentification, obtenue en utilisant la fonction \code{api_connect}.}

\item{start_date}{La date de début de la période de sélection des données (format "yyyy-mm-dd").}

\item{end_date}{La date de fin de la période de sélection des données (format "yyyy-mm-dd").}

\item{query}{Requête pour filtrer les données par mots-clés (facultatif), ex: query='securité au sahel'.}

\item{source}{Source des données (facultatif), ex: source=c('journaldumali.com', 'faso.net').}

\item{category}{Catégorie des données (facultatif), ex: category=c('economique','politique').}

\item{country}{Pays des données (facultatif), ex: country=c('ml')}
}
\value{
Une table de données, avec leurs détails tels que l'identifiant, le titre, la description, le contenu, la date de publication, le pays, le tri par pays, le nombre de commentaires et les mentions.
}
\description{
Cette fonction récupère les données publiées dans les variables de filtrage configurées. Pour plus de détails, consultez la documentation à l'adresse suivante : https://docs.newswacafi.online/
}
\details{
Possible endpoint values:

\begin{table}
\centering
\begin{tabular}{|l|}
\hline
\textbf{Source} \\ \hline
lefaso.net \\
burkina24.com \\
sidwaya.info \\
aib.media \\
lesahel.org \\
actuniger.com \\
levenementniger.com \\
journaldumali.com \\
maliweb.net \\
maliactu.net \\ \hline
\end{tabular}
\caption{Possible sources}
\label{tab:sources}
\end{table}

\begin{table}
\centering
\begin{tabular}{|l|}
\hline
\textbf{Category} \\ \hline
Politics \\
Economics \\
Society \\
Governance \\ \hline
\end{tabular}
\caption{Possible categories}
\label{tab:categories}
\end{table}

\begin{table}
\centering
\begin{tabular}{|l|l|}
\hline
\textbf{Country} & \textbf{Country code} \\ \hline
Mali & ml \\
Burkina Faso & bf \\
Niger & ne \\ \hline
\end{tabular}
\caption{Possible countries and country codes}
\label{tab:countries-codes}
\end{table}
}
\examples{

# Utilisation de la fonction pour récupérer les news par date
api_key <- "8c2cf1af-c970-47dd-a3fb-9841b2e9e8ed"
api_secret <- "pot_SvtyFegiqK2YckFNfAXPgjdNtVrM8BdW_jKd-zo"
connect_info <- api_connect(api_key = api_key, api_secret = api_secret)
start_date <- as.Date("2024-01-01")
end_date <- as.Date("2024-01-31")
news <- get_news(connect_info, start_date, end_date)
head(news,2)

}
