% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{api_connect}
\alias{api_connect}
\title{Fonction pour établir la connexion à l'API}
\usage{
api_connect(api_key, api_secret)
}
\arguments{
\item{api_key}{La clé d'API utilisée pour l'authentification.}

\item{api_secret}{Le secret d'API utilisé pour l'authentification.}
}
\value{
Une liste contenant l'URL de base de l'API et les en-têtes d'authentification.
}
\description{
Cette fonction établit une connexion à l'API en utilisant les clés d'API fournies.
}
\examples{
# Utilisation de la fonction pour établir une connexion à l'API
api_key <- "6a07a7ee-ade3-4209-aa1b-b4504ebc0e68"
api_secret <- "pbkdf2_sha256$600000$eVJKaltgn341HcwJWfVlWg$bpuIG09itt+g1R2MDOzS/QMoJstT6bWZTTVTcQlQ3cc="
connection_info <- api_connect(api_key = api_key, api_secret = api_secret)

}
